// src/types/generated/ipc/RecordToolInvocationInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ToolEffectClass } from "./ToolEffectClass";
import type { JsonValue } from "./serde_json/JsonValue";

export type RecordToolInvocationInput = { toolName: string, effectClass: ToolEffectClass, arguments: JsonValue, remoteExecutionId: string | null, conversationId: string, turnId: string, messageId: string, callId: string, };